requirements.txt
cloudpickle==1.6.0
httpx==0.21.1
numpy==1.21.4
Pillow==8.4.0
pandas==1.2.3
pytest==6.2.5
networkx==2.7.1

-f https://download.pytorch.org/whl/lts/1.8/torch_lts.html
torch==1.8.2+cpu